feat: cleanup-ui-worker-editor-section-boundaries-001 - Preserve Ready-State Shell Behavior While Delegating Fields
