+++ ui/src/features/current-selection/worker-selection/components/editable/fields/worker-editable-configuration-hosted-fields.tsx
@@ -1,0 +1,63 @@
+import { EDITABLE_HOSTED_PROVIDERS } from "../../../../../current-factory-definition/lib/worker-editable-values";
+import type {
+  ReadyWorkerEditableConfigurationState,
+  ReadyWorkerEditableValidationErrors,
+  WorkerEditableConfigurationMessages,
+} from "./worker-editable-configuration-field-types";
+import {
+  WorkerEditableConfigurationField,
+  WorkerEditableConfigurationServerChangedHint,
+  WorkerOptionalEnumSelect,
+} from "./worker-editable-configuration-field-primitives";
+import { LinearHostedWorkerEditableFields } from "./worker-editable-configuration-linear-hosted-fields";
+
+export function WorkerEditableConfigurationHostedFields({
+  messages,
+  state,
+  validationErrors,
+}: {
+  messages: WorkerEditableConfigurationMessages;
+  state: ReadyWorkerEditableConfigurationState;
+  validationErrors: ReadyWorkerEditableValidationErrors;
+}) {
+  return (
+    <>
+      <WorkerEditableConfigurationField
+        errorMessage={validationErrors.provider}
+        fieldId="editable-worker-provider"
+        input={
+          <WorkerOptionalEnumSelect
+            ariaDescribedBy={
+              validationErrors.provider
+                ? "editable-worker-provider-error"
+                : undefined
+            }
+            ariaInvalid={Boolean(validationErrors.provider)}
+            id="editable-worker-provider"
+            label={messages.providerFieldLabel}
+            notConfiguredLabel={messages.notConfiguredOptionLabel}
+            onChange={state.onProviderChange}
+            options={EDITABLE_HOSTED_PROVIDERS}
+            renderLabel={(value) => value}
+            value={state.draft.provider}
+          />
+        }
+        label={messages.providerFieldLabel}
+        supportingContent={
+          <WorkerEditableConfigurationServerChangedHint
+            fieldName="provider"
+            messages={messages}
+            state={state}
+          />
+        }
+      />
+      {state.draft.provider === "LINEAR" ? (
+        <LinearHostedWorkerEditableFields
+          messages={messages}
+          state={state}
+          validationErrors={validationErrors}
+        />
+      ) : null}
+    </>
+  );
+}
