+++ ui/src/features/current-selection/worker-selection/components/editable/worker-editable-configuration-section.checkbox-verification.stories.tsx
@@ -2,9 +2,9 @@ import { expect, userEvent, within } from "storybook/test";
 import { useState } from "react";
 
 import "../../../../styles.css";
-import type { EditableWorkerSaveValidationErrors } from "../lib/detail-card-types";
-import { getWorkerDetailMessages } from "../messages/worker-detail";
-import { expectStyledCheckboxInStory } from "../../../../testing/checkbox-story-helpers";
+import type { EditableWorkerSaveValidationErrors } from "../../lib/detail-card-types";
+import { getWorkerDetailMessages } from "../../messages/worker-detail";
+import { expectStyledCheckboxInStory } from "../../../../../testing/checkbox-story-helpers";
 import { WorkerEditableConfigurationSection } from "./worker-editable-configuration-section";
 
 const messages = getWorkerDetailMessages();
