+++ ui/src/features/current-selection/worker-selection/public/index.ts
@@ -2,4 +2,4 @@ export { WorkerDetailCard } from "../components/worker-detail-card";
 export {
   EditableWorkerConfigurationHeaderActions,
   EditableWorkerSaveHeaderAction,
-} from "../components/worker-save-controls";
+} from "../components/editable/worker-save-controls";
